+++ android/src/main/kotlin/com/adyen/checkout/flutter/dropIn/model/DropInServiceEvent.kt
@@ -1,0 +1,5 @@
+package com.adyen.checkout.flutter.dropIn.model
+
+enum class DropInServiceEvent {
+    STOP
+}