+++ lib/src/common/adyen_checkout_advanced.dart
@@ -32,4 +32,6 @@ class AdyenCheckoutAdvanced {
         configuration,
         paymentMethod,
       );
+
+  Future<void> stopDropIn() async => await dropIn.stopDropIn();
 }