+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### New
 
-- Added bin lookup callbacks for Card Component and Drop-In. 
+- Added bin lookup callbacks for Card Component and Drop-In.
+- Added custom card example using CSE and card validators.
 - Set minimum SDK version to Flutter 3.16/Dart 3.2
 - Android Components/Drop-in
   version: [5.9.1](https://docs.adyen.com/online-payments/release-notes/?title%5B0%5D=Android+Components%2FDrop-in#releaseNote=2025-02-24-android-componentsdrop-in-5.9.1).